+++ docs/changes.rst
@@ -9,6 +9,8 @@ Changes
 TODO
 ====
 
+* In Logya create: copy content from base site first and then selected theme.
+* Rename site to theme.
 * Rename site option to theme or remove option and sites dir to themes. The current option name site is less clear.
 * sitemap.xml directory is generated in public instead of file.
 * Implement noindex attribute for docs.